refactor: Move common helpers to lib
extract config getter and combine to single file
