+++ properties/dependencies.js
@@ -1,4 +1,4 @@
-const readDeps = require('../lib/read-deps')
+const { readDeps } = require('../lib')
 
 module.exports = ({ config, dirname, package: pkg }) => {
   return readDeps(config, dirname, false, pkg.devDependencies || {})
